Move device factory registry to another module
